completed 4.21

+++ controllers/blogs.js
@@ -37,8 +37,20 @@ blogsRouter.post('/', async (request, response)=>{
 })
 
 blogsRouter.delete('/:id', async(request, response)=>{
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id){
+    return response.status(401).json({error: 'token missing or invaild'})
+  } 
+
+  const blog = await Blog.findById(request.params.id)
+  if(blog.user.toString() === decodedToken.id.toString()){
+    console.log(blog)
+    console.log(decodedToken)
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else{
+    return response.status(401).json({error: 'cannot delete blog from different user'})
+  }
 })
 
 blogsRouter.put('/:id', async(request, response)=>{
